refactor(keepcommon): flatten cache, encryption, logging, persistence, rate, wrappers, generate, chain/local to pkg/
These packages have zero naming collisions with existing pkg/ packages
and no codegen-tool entanglement, so they move out of the keepcommon/
prefix mechanically.
