+++ pkg/beacon/event/deduplicator.go
@@ -3,7 +3,7 @@ package event
 import (
 	"encoding/hex"
 	"fmt"
-	"github.com/keep-network/keep-core/pkg/keepcommon/cache"
+	"github.com/keep-network/keep-core/pkg/cache"
 	"math/big"
 	"sync"
 	"time"
