+++ pkg/crypto/ephemeral/symmetric_key.go
@@ -4,7 +4,7 @@ import (
 	"crypto/sha256"
 
 	"github.com/btcsuite/btcd/btcec"
-	"github.com/keep-network/keep-core/pkg/keepcommon/encryption"
+	"github.com/keep-network/keep-core/pkg/encryption"
 )
 
 // SymmetricEcdhKey is an ephemeral Elliptic Curve key created with
