+++ pkg/firewall/firewall_test.go
@@ -7,9 +7,9 @@ import (
 	"time"
 
 	"github.com/keep-network/keep-core/internal/testutils"
+	"github.com/keep-network/keep-core/pkg/cache"
 	"github.com/keep-network/keep-core/pkg/chain/local_v1"
 	"github.com/keep-network/keep-core/pkg/clientinfo"
-	"github.com/keep-network/keep-core/pkg/keepcommon/cache"
 	"github.com/keep-network/keep-core/pkg/operator"
 )
