+++ pkg/keepcommon/chain/ethereum/ethutil/rate_limiter.go
@@ -8,7 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/keep-network/keep-core/pkg/keepcommon/rate"
+	"github.com/keep-network/keep-core/pkg/rate"
 )
 
 type rateLimiter struct {
