+++ pkg/persistence/encrypted_persistence_test.go
@@ -7,7 +7,7 @@ import (
 
 	"crypto/sha256"
 
-	"github.com/keep-network/keep-core/pkg/keepcommon/encryption"
+	"github.com/keep-network/keep-core/pkg/encryption"
 )
 
 const accountPassword = "grzeski"
