+++ pkg/tbtc/node.go
@@ -11,8 +11,8 @@ import (
 	"github.com/keep-network/keep-core/pkg/clientinfo"
 
 	"github.com/keep-network/keep-core/pkg/generator"
-	"github.com/keep-network/keep-core/pkg/keepcommon/persistence"
 	"github.com/keep-network/keep-core/pkg/net"
+	"github.com/keep-network/keep-core/pkg/persistence"
 )
 
 const (
